--- app/(tabs)/showcar.tsx.old
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Showcar = () => {
-    return (
-        <div>Showcar</div>
-    )
-}
-
-export default Showcar
